azure-functions/requirements.txt
azure-functions
requests
pandas
numpy
